feat: change quantity type from Double to Integer in product and order models, update related classes and database schema

+++ services/notification/src/main/java/com/mclavo/ecommerce/order/Product.java
@@ -7,6 +7,6 @@ public record Product(
     String name,
     String description,
     BigDecimal price,
-    Double quantity
+    Integer quantity
 
 ) {}
